Add example workflow options for infer and flawfinder, fix defaults, sh
In the C++ workflow eg in the README.md, examples for fbinfer_args,
 flawfinder_args, and flawfinder_targets are now present that I added.
Fixed defaults for cmake_args, flawfinder_targets, and fbinfer_args in
 the action.yml and README.

Preprocess fbinfer_args, flawfinder_args, and flawfinder_targets in
 line with the way cppcheck and clang_tidy arg preprocessing is
 handled.

Add -fi and -ff options to python3 -m src.static_analysis_cpp commands,
 note that I have not yet added these to the python code yet.
-fi receives the file name for the infer report, and -ff receives the
 file name for the flawfinder report.

Output the values of INFER_ARGS, FLAWFINDER_ARGS, and FLAWFINDER_TGTS
 for debugging where necessary.

Added calls to infer for CMake and non-CMake sections.
Add call to flawfinder, that does not depend at all on whether CMake is
 being used.

Ordered the variables and operations so everything is consistent.